reduce post tag form to simple toggle handler
