+++ src/components/ApplicationViews.js
@@ -31,9 +31,11 @@ export const ApplicationViews = () => {
       <main className="container p-5">
         <PostProvider>
           <CategoryProvider>
-            <Route path="/posts/create" component={PostForm} />
+            <TagProvider>
+              <Route path="/posts/create" component={PostForm} />
 
-            <Route path="/posts/edit/:postId" component={PostForm} />
+              <Route path="/posts/edit/:postId" component={PostForm} />
+            </TagProvider>
           </CategoryProvider>
         </PostProvider>
         <PostProvider>
